refactor: separate components into different files
