+++ lib/presentation/onboarding/preparation_name_select/components/preparation_select_list.dart
@@ -1,13 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:on_time_front/presentation/onboarding/components/onboarding_page_view_layout.dart';
+import 'package:on_time_front/presentation/onboarding/preparation_name_select/components/preparation_name_select_field.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_name_select/cubit/preparation_name/preparation_name_cubit.dart';
 import 'package:on_time_front/presentation/onboarding/preparation_name_select/cubit/preparation_step_name/preparation_step_name_cubit.dart';
-import 'package:on_time_front/presentation/shared/components/check_button.dart';
-import 'package:on_time_front/presentation/shared/components/tile.dart';
-import 'package:on_time_front/presentation/shared/theme/tile_style.dart';
-
-typedef OnSelectedStepChangedCallBackFunction<T> = void Function(List<T>);
 
 class PreparationSelectList extends StatelessWidget {
   const PreparationSelectList({super.key});
@@ -99,119 +94,5 @@ class PreparationSelectList extends StatelessWidget {
         ),
       );
     });
-  }
-}
-
-class PreparationNameSelectField extends StatefulWidget {
-  const PreparationNameSelectField({
-    super.key,
-    required this.preparationStep,
-    this.onNameChanged,
-    this.onNameSaved,
-    required this.onSelectionChanged,
-    this.isAdding = false,
-  });
-
-  final PreparationStepNameState preparationStep;
-  final ValueChanged<String>? onNameChanged;
-  final VoidCallback? onNameSaved;
-  final VoidCallback onSelectionChanged;
-  final bool isAdding;
-
-  @override
-  State<PreparationNameSelectField> createState() =>
-      _PreparationNameSelectFieldState();
-}
-
-class _PreparationNameSelectFieldState
-    extends State<PreparationNameSelectField> {
-  final FocusNode focusNode = FocusNode();
-
-  @override
-  void dispose() {
-    focusNode.dispose();
-    super.dispose();
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    focusNode.addListener(() {
-      if (!focusNode.hasFocus) {
-        widget.onNameSaved?.call();
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (widget.isAdding) {
-      focusNode.requestFocus();
-    }
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 8.0),
-      child: Tile(
-        key: ValueKey<String>(widget.preparationStep.preparationId),
-        style: TileStyle(padding: EdgeInsets.all(16.0)),
-        leading: SizedBox(
-            width: 30,
-            height: 30,
-            child: CheckButton(
-              isChecked: widget.preparationStep.isSelected,
-              onPressed: widget.onSelectionChanged,
-            )),
-        child: Expanded(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 19.0),
-            child: Container(
-              constraints: BoxConstraints(minHeight: 30),
-              child: TextFormField(
-                initialValue: widget.preparationStep.preparationName.value,
-                onChanged: widget.onNameChanged,
-                onFieldSubmitted: (value) => widget.onNameSaved?.call(),
-                onTapOutside: (event) {
-                  FocusManager.instance.primaryFocus?.unfocus();
-                },
-                decoration: InputDecoration(
-                  isDense: true,
-                  border: InputBorder.none,
-                  contentPadding: EdgeInsets.all(3.0),
-                ),
-                focusNode: focusNode,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class PreparationSelectField extends StatefulWidget {
-  const PreparationSelectField({
-    super.key,
-  });
-
-  @override
-  State<PreparationSelectField> createState() => _PreparationSelectFieldState();
-}
-
-class _PreparationSelectFieldState extends State<PreparationSelectField> {
-  @override
-  void initState() {
-    super.initState();
-    context.read<PreparationNameCubit>().initialize();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
-    return OnboardingPageViewLayout(
-      title: Text(
-        '주로 하는 준비 과정을\n선택해주세요',
-        style: textTheme.titleLarge,
-      ),
-      child: PreparationSelectList(),
-    );
   }
 }
